Show Metals treeview associated with current buffer filename.
Fix a bug with multiple workspaces, when closing one workspace
the treeview will be closed. When a file associated with the closed
workspace was open and invoking lsp-metals-treeview the other
workspace's treeview would be displayed.
lsp-metals-treeview now uses lsp-find-workspace with metals
server-id and filename of current buffer to locate the
associated workspace.